Rewrite how ultimates work.

+++ src/main/java/me/hapyl/fight/util/displayfield/DisplayField.java
@@ -55,6 +55,6 @@ public @interface DisplayField {
      *
      * @return the decimal point.
      */
-    double dp() default -1;
+    int dp() default -1;
 
 }
